+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* **2026-02-19**: v0.2.5 - v0.2.5 - fix internationalization (Hannibal)   
+
+* **2026-02-20**: v0.3.6 - nano, simple and rich brief (Hannibal)   
 
 * **2026-02-19**: v0.2.4 - general fixes and query extraction from assistant message (Hannibal)